Delete test file and add tests to test_utils.py
